2019-12-08 Roman (stable) - !!:FIX:SYS: The JavaLikeCalc translation by the function tr() fixed in creating the translations table "Trs" and entries here in     the simple mode (single language).     !!:FIX:The function TSYS::chkSelDB() fixed after expanding for checking the DB enabling, checking the special DBs DB_CFG, "*.*". - FIX:Archive: TArchiveS::rdRestDtOverTm() fixed by expanding from the float to double type,     due to the precission errors at multiplication on 86400 in reloading from the local archive of the redundancy. - !!!!:SYS.Transport,Archive.FSArch,DAQ.{BFN,GPIO,ModBus,OPC_UA},Protocol.{OPC_UA,HTTP,SelfSystem},UI.{VCAEngine,Vision,WebVision}:     Some common data resources of dataRes() replaced by own resources, mostly due to the observed locking problems in Protocol.HTTP, DAQ.ModBus. - DB.MySQL: The default timeouts set to "10:5:5". - !!: DB.{MySQL,FireBird,PostgreSQL}: The connection restoring appended by limit in 3 tries, to prevent of hangs at timeout errors.     Tables list obtaining of the DB switched to the indiferent mode to transactions. - DAQ.ModBus: Appended for writing into the logical context in the redundant mode.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2654 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tarchives.h
@@ -245,8 +245,8 @@ class TArchiveS : public TSubSYS
 	// Redundancy
 	time_t rdTm( );			//Time of the last redundancy operation
 	bool rdProcess( XMLNode *reqSt = NULL );
-	float rdRestDtOverTm( )			{ return mRdRestDtOverTm; }	//In days
-	void setRdRestDtOverTm( float vl )	{ mRdRestDtOverTm = vmin(356,vmax(0,vl)); modif(); }
+	double rdRestDtOverTm( )		{ return mRdRestDtOverTm; }	//In days
+	void setRdRestDtOverTm( double vl )	{ mRdRestDtOverTm = vmin(356, vmax(0,vl)); modif(); }
 	void rdActArchMList( vector<string> &ls, bool isRun = false );
 	string rdStRequest( const string &arch, XMLNode &req, const string &prevSt = "", bool toRun = true );
 
@@ -306,7 +306,7 @@ class TArchiveS : public TSubSYS
 
 	// Redundancy
 	ResRW	mRdRes;
-	float	mRdRestDtOverTm;	//Overtime of the redundant history reload at start in hours
+	double	mRdRestDtOverTm;	//Overtime of the redundant history reload at start in hours
 	map<string, map<string,bool> > mRdArchM;
 	unsigned mRdFirst	: 1;
 };
